[FAM] - chore: package for fam generator (#1020)
