+++ src/apps/Loop/components/Sidesheet/Checklists.tsx
@@ -1,3 +1,4 @@
+import { generateExpressions, generateFamRequest } from '@equinor/fam-request-builder';
 import { statusColorMap, TabTable } from '@equinor/GardenUtils';
 import {
     CellProps,
@@ -9,7 +10,6 @@ import { useQuery } from 'react-query';
 import { proCoSysUrls } from '../../../../packages/ProCoSysUrls/procosysUrl';
 import { ChecklistForLoop } from '../../types';
 import { checklistColumnNames, getChecklistsForLoop } from '../../utility/api';
-import { generateExpressions, generateFamRequest } from '../../utility/helpers/fam';
 
 const columns: Column<ChecklistForLoop>[] = [
     {
