+++ src/apps/ScopeChangeRequest/api/FAM/getPunchListItemByNo.ts
@@ -1,5 +1,5 @@
+import { generateExpressions, generateFamRequest } from '@equinor/fam-request-builder';
 import { httpClient } from '@equinor/lighthouse-portal-client';
-import { generateExpressions, generateFamRequest } from '../../functions/FAM/generateFAMRequest';
 import { throwOnError } from '../../functions/throwError';
 import { PunchListItem } from '../../types/FAM/punchListItem';
