+++ tsconfig.json
@@ -145,6 +145,9 @@
             ],
              "@equinor/modal": [
                 "./src/packages/Modal"
+            ],
+            "@equinor/fam-request-builder": [
+                "./src/packages/FamRequestBuilder"
             ]
         }
     },
